Remove `helpers.js`.
