+++ site/assets/js/pages/checkout/form-constants.js
@@ -1,0 +1,33 @@
+'use strict';
+
+export const requiredSelector = 'input[required], select[required], textarea[required]';
+
+export const countryPhoneCodes = {
+    AT: '43',
+    BE: '32',
+    BG: '359',
+    HR: '385',
+    CY: '357',
+    CZ: '420',
+    DK: '45',
+    EE: '372',
+    FI: '358',
+    FR: '33',
+    DE: '49',
+    GR: '30',
+    HU: '36',
+    IE: '353',
+    IT: '39',
+    LV: '371',
+    LT: '370',
+    LU: '352',
+    MT: '356',
+    NL: '31',
+    PL: '48',
+    PT: '351',
+    RO: '40',
+    SK: '421',
+    SI: '386',
+    ES: '34',
+    SE: '46'
+};
